separate rename module
